pyproject.toml
[tool.black]
line-length = 79
exclude = '''
/(
    \.venv
)/
'''

[tool.isort]
multi_line_output=3
include_trailing_comma = true

[tool.poetry]
name = "kahless"
version = "0.1"
description = ""
license = "BSD-3-Clause"
authors = ["Lech Gudalewicz <lechgu@gmail.com>"]
repository = "https://github.com/lechgu/kahless"

[tool.poetry.dependencies]
python = "~3.7 || ~3.8"
click = "^7.0"
gunicorn = "^20.0.4"
webob = "^1.8.6"
parse = "^1.15.0"
requests = "^2.23.0"
requests-wsgi-adapter = "^0.4.1"
jinja2 = "^2.11.2"
whitenoise = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.7.9"
pep8-naming = "^0.9.1"
flake8-isort = "^2.8.0"
pycalver = "^201903.30"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
